+++ packages/asc-ui/src/components/Paragraph/ParagraphStyle.ts
@@ -1,14 +1,17 @@
 import styled, { css } from 'styled-components'
 import type { Props as TypographyProps } from '../Typography/TypographyStyle'
 import TypographyStyle from '../Typography/TypographyStyle'
+import { calculateFluidStyle } from '../../utils'
 
-export type Props = TypographyProps
-
-export const ParagraphStyleCSS = css<Props>`
-  margin-top: 0;
-  color: inherit;
-`
+export type Props = {
+  intro?: boolean
+} & TypographyProps
 
 export default styled(TypographyStyle)<Props>`
-  ${ParagraphStyleCSS};
+  ${({ intro }) =>
+    intro &&
+    css`
+      font-size: ${calculateFluidStyle(22, 30)};
+      line-height: ${calculateFluidStyle(35, 45)};
+    `}
 `